Rename title HTML helper for clarity
