+++ capy/src/main/java/com/jocmp/capy/accounts/feedbin/FeedbinAccountDelegate.kt
@@ -11,7 +11,7 @@ import com.jocmp.capy.accounts.withErrorHandling
 import com.jocmp.capy.common.TimeHelpers
 import com.jocmp.capy.common.UnauthorizedError
 import com.jocmp.capy.common.host
-import com.jocmp.capy.common.strippedOfTitleHTML
+import com.jocmp.capy.common.stripTitleMarkup
 import com.jocmp.capy.common.toDateTime
 import com.jocmp.capy.common.toDateTimeFromSeconds
 import com.jocmp.capy.common.transactionWithErrorHandling
@@ -458,7 +458,7 @@ internal class FeedbinAccountDelegate(
                 database.articlesQueries.create(
                     id = articleID,
                     feed_id = entry.feed_id.toString(),
-                    title = entry.title?.strippedOfTitleHTML,
+                    title = entry.title?.stripTitleMarkup(),
                     author = entry.author,
                     content_html = entry.content,
                     extracted_content_url = entry.extracted_content_url,
