+++ capy/src/main/java/com/jocmp/capy/accounts/miniflux/MinifluxAccountDelegate.kt
@@ -9,7 +9,7 @@ import com.jocmp.capy.accounts.withErrorHandling
 import com.jocmp.capy.common.ContentFormatter
 import com.jocmp.capy.common.TimeHelpers
 import com.jocmp.capy.common.UnauthorizedError
-import com.jocmp.capy.common.strippedOfTitleHTML
+import com.jocmp.capy.common.stripTitleMarkup
 import com.jocmp.capy.common.toDateTime
 import com.jocmp.capy.common.transactionWithErrorHandling
 import com.jocmp.capy.db.Database
@@ -395,7 +395,7 @@ internal class MinifluxAccountDelegate(
                 database.articlesQueries.create(
                     id = articleID,
                     feed_id = entry.feed_id.toString(),
-                    title = entry.title.strippedOfTitleHTML,
+                    title = entry.title.stripTitleMarkup(),
                     author = entry.author,
                     content_html = entry.content,
                     extracted_content_url = null,
